(815) Redirect submitted referral redux

+++ server/controllers/refer/courseParticipationsController.ts
@@ -16,7 +16,13 @@ export default class CourseParticipationsController {
     return async (req: Request, res: Response) => {
       TypeUtils.assertHasUser(req)
 
-      const referral = await this.referralService.getReferral(req.user.token, req.params.referralId)
+      const { referralId } = req.params
+
+      const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
 
       const { courseName, hasFormErrors } = CourseParticipationUtils.processCourseFormData(
         req.body.courseName,
@@ -25,7 +31,7 @@ export default class CourseParticipationsController {
       )
 
       if (hasFormErrors) {
-        return res.redirect(referPaths.programmeHistory.new({ referralId: req.params.referralId }))
+        return res.redirect(referPaths.programmeHistory.new({ referralId }))
       }
 
       const courseParticipation = await this.courseService.createParticipation(
@@ -51,12 +57,17 @@ export default class CourseParticipationsController {
 
       const { courseParticipationId, referralId } = req.params
 
+      const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
+
       const courseParticipation = await this.courseService.getParticipation(req.user.token, courseParticipationId)
       const courseParticipationPresenter = await this.courseService.presentCourseParticipation(
         req.user.token,
         courseParticipation,
       )
-      const referral = await this.referralService.getReferral(req.user.token, referralId)
       const person = await this.personService.getPerson(
         req.user.username,
         referral.prisonNumber,
@@ -68,11 +79,11 @@ export default class CourseParticipationsController {
         remove: false,
       })
 
-      res.render('referrals/courseParticipations/delete', {
+      return res.render('referrals/courseParticipations/delete', {
         action: `${referPaths.programmeHistory.destroy({ courseParticipationId, referralId })}?_method=DELETE`,
         pageHeading: 'Remove programme',
         person,
-        referralId: referral.id,
+        referralId,
         summaryListOptions,
       })
     }
@@ -83,6 +94,12 @@ export default class CourseParticipationsController {
       TypeUtils.assertHasUser(req)
 
       const { courseParticipationId, referralId } = req.params
+
+      const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
 
       await this.courseService.deleteParticipation(req.user.token, courseParticipationId)
 
@@ -102,8 +119,13 @@ export default class CourseParticipationsController {
 
       const { courseParticipationId, referralId } = req.params
 
-      const courseParticipation = await this.courseService.getParticipation(req.user.token, courseParticipationId)
       const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
+
+      const courseParticipation = await this.courseService.getParticipation(req.user.token, courseParticipationId)
       const person = await this.personService.getPerson(
         req.user.username,
         referral.prisonNumber,
@@ -118,7 +140,7 @@ export default class CourseParticipationsController {
       const formValues = isKnownCourse ? { courseName } : { courseName: 'Other', otherCourseName: courseName }
       const hasOtherCourseNameError = !!res.locals.errors.messages.otherCourseName
 
-      res.render('referrals/courseParticipations/course', {
+      return res.render('referrals/courseParticipations/course', {
         action: `${referPaths.programmeHistory.updateProgramme({ courseParticipationId, referralId })}?_method=PUT`,
         courseRadioOptions: CourseUtils.courseRadioOptions(courseNames),
         formValues,
@@ -134,8 +156,14 @@ export default class CourseParticipationsController {
     return async (req: Request, res: Response) => {
       TypeUtils.assertHasUser(req)
 
-      const successMessage = req.flash('successMessage')[0]
-      const referral = await this.referralService.getReferral(req.user.token, req.params.referralId)
+      const { referralId } = req.params
+
+      const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
+
       const person = await this.personService.getPerson(
         req.user.username,
         referral.prisonNumber,
@@ -152,14 +180,16 @@ export default class CourseParticipationsController {
       )
 
       const summaryListsOptions = courseParticipationsPresenter.map(participation =>
-        CourseParticipationUtils.summaryListOptions(participation, referral.id),
+        CourseParticipationUtils.summaryListOptions(participation, referralId),
       )
 
-      res.render('referrals/courseParticipations/index', {
+      const successMessage = req.flash('successMessage')[0]
+
+      return res.render('referrals/courseParticipations/index', {
         action: `${referPaths.programmeHistory.updateReviewedStatus({ referralId: referral.id })}?_method=PUT`,
         pageHeading: 'Accredited Programme history',
         person,
-        referralId: referral.id,
+        referralId,
         successMessage,
         summaryListsOptions,
       })
@@ -169,9 +199,15 @@ export default class CourseParticipationsController {
   new(): TypedRequestHandler<Request, Response> {
     return async (req: Request, res: Response) => {
       TypeUtils.assertHasUser(req)
+      const { referralId } = req.params
+
+      const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
 
       const courseNames = await this.courseService.getCourseNames(req.user.token)
-      const referral = await this.referralService.getReferral(req.user.token, req.params.referralId)
       const person = await this.personService.getPerson(
         req.user.username,
         referral.prisonNumber,
@@ -180,8 +216,8 @@ export default class CourseParticipationsController {
 
       FormUtils.setFieldErrors(req, res, ['courseName', 'otherCourseName'])
 
-      res.render('referrals/courseParticipations/course', {
-        action: referPaths.programmeHistory.create({ referralId: referral.id }),
+      return res.render('referrals/courseParticipations/course', {
+        action: referPaths.programmeHistory.create({ referralId }),
         courseRadioOptions: CourseUtils.courseRadioOptions(courseNames),
         formValues: {},
         otherCourseNameChecked: !!res.locals.errors.messages.otherCourseName,
@@ -196,9 +232,17 @@ export default class CourseParticipationsController {
     return async (req: Request, res: Response) => {
       TypeUtils.assertHasUser(req)
 
+      const { courseParticipationId, referralId } = req.params
+
+      const referral = await this.referralService.getReferral(req.user.token, referralId)
+
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
+
       const currentCourseParticipation = await this.courseService.getParticipation(
         req.user.token,
-        req.params.courseParticipationId,
+        courseParticipationId,
       )
 
       const { courseName, hasFormErrors } = CourseParticipationUtils.processCourseFormData(
@@ -210,15 +254,15 @@ export default class CourseParticipationsController {
       if (hasFormErrors) {
         return res.redirect(
           referPaths.programmeHistory.editProgramme({
-            courseParticipationId: currentCourseParticipation.id,
-            referralId: req.params.referralId,
+            courseParticipationId,
+            referralId,
           }),
         )
       }
 
       const { detail, outcome, setting, source } = currentCourseParticipation
 
-      await this.courseService.updateParticipation(req.user.token, currentCourseParticipation.id, {
+      await this.courseService.updateParticipation(req.user.token, courseParticipationId, {
         courseName: courseName as CourseParticipation['courseName'],
         detail,
         outcome,
@@ -230,8 +274,8 @@ export default class CourseParticipationsController {
 
       return res.redirect(
         referPaths.programmeHistory.details.show({
-          courseParticipationId: req.params.courseParticipationId,
-          referralId: req.params.referralId,
+          courseParticipationId,
+          referralId,
         }),
       )
     }
@@ -243,9 +287,14 @@ export default class CourseParticipationsController {
 
       const { referralId } = req.params
 
-      const hasReviewedProgrammeHistory = req.body.hasReviewedProgrammeHistory === 'true'
+      const referral = await this.referralService.getReferral(req.user.token, req.params.referralId)
 
-      const { oasysConfirmed, reason } = await this.referralService.getReferral(req.user.token, referralId)
+      if (referral.status !== 'referral_started') {
+        return res.redirect(referPaths.complete({ referralId }))
+      }
+
+      const hasReviewedProgrammeHistory = req.body.hasReviewedProgrammeHistory === 'true'
+      const { oasysConfirmed, reason } = referral
 
       const referralUpdate: ReferralUpdate = {
         hasReviewedProgrammeHistory,
